modify mechanic pipeline
