+++ pipeline/code_gen/gen_mechanic/contracts.py
@@ -1,0 +1,80 @@
+"""Mechanic contract schema and deterministic validation."""
+
+from __future__ import annotations
+
+from collections.abc import Mapping, Sequence
+from pathlib import Path
+from typing import Any
+
+from pipeline.common.artifacts import read_json
+
+
+MECHANIC_CONTRACT_SCHEMA = (
+    "aaagameforge.mechanic_contract.v1"
+)
+MECHANIC_CONTRACT_FILENAME = "mechanic_contract.json"
+
+
+def _contract_collection_present(value: Any) -> bool:
+    if isinstance(value, Mapping):
+        return bool(value)
+    if isinstance(value, Sequence) and not isinstance(
+        value,
+        (str, bytes),
+    ):
+        return bool(value)
+    return False
+
+
+def validate_mechanic_contract(
+    workspace: Path,
+    gameplay_module_name: str,
+) -> tuple[bool, list[str]]:
+    contract_path = workspace / MECHANIC_CONTRACT_FILENAME
+    if not contract_path.is_file():
+        return False, [
+            "Required Mechanic contract is missing: "
+            f"{MECHANIC_CONTRACT_FILENAME}"
+        ]
+    try:
+        contract = read_json(
+            contract_path,
+            "Mechanic contract",
+        )
+    except (FileNotFoundError, TypeError, ValueError) as exc:
+        return False, [str(exc)]
+
+    errors: list[str] = []
+    if contract.get("schema_version") != (
+        MECHANIC_CONTRACT_SCHEMA
+    ):
+        errors.append(
+            "Mechanic contract schema_version must be "
+            f"{MECHANIC_CONTRACT_SCHEMA!r}"
+        )
+    contract_version = contract.get("contract_version")
+    if (
+        not isinstance(contract_version, int)
+        or isinstance(contract_version, bool)
+        or contract_version <= 0
+    ):
+        errors.append(
+            "Mechanic contract contract_version must be a "
+            "positive integer"
+        )
+    if str(contract.get("gameplay_module") or "") != (
+        gameplay_module_name
+    ):
+        errors.append(
+            "Mechanic contract gameplay_module must match "
+            f"{gameplay_module_name!r}"
+        )
+    for section in ("state", "events", "commands"):
+        if not _contract_collection_present(
+            contract.get(section)
+        ):
+            errors.append(
+                "Mechanic contract must define a non-empty "
+                f"{section} collection"
+            )
+    return not errors, errors
